--- a/.gitignore
+++ b/.gitignore
@@ -207,3 +207,6 @@ images/
 *.exp
 *.lib
 *.obj
+
+# generated files
+**/generated/**