--- a/aphrodite/quantization/utils/__init__.py
+++ b/aphrodite/quantization/utils/__init__.py
@@ -1,0 +1,3 @@
+from .layer_utils import replace_parameter, update_tensor_inplace
+
+__all__ = ['update_tensor_inplace', 'replace_parameter']