TL;DR: 46% throughput increase for GPTQ models on Hopper GPUs.

This PR adds the Machete kernels, spiritual successor to the Marlin kernels, developed by Neural Magic.

Marlin uses `mma` instructions, which are the fastest tensor core instructions you can use on Ampere GPUs. Hopper however has `wgmma`, which are necessary for achieving peak FLOPs on Hopper GPUs, and without them (on `mma`), you're expected to only hit ~75% of peak. See [ThunderKittens](https://hazyresearch.stanford.edu/blog/2024-05-12-tk) and [Benchmarking and Dissecting the Nvidia Hopper GPU Architecture](https://arxiv.org/pdf/2402.13499).

Marlin uses a specific weight storage layout that is specialized for the `mma` instructions, so we will want to adopt a more flexible way of defining these layouts, so we can accommodate new instructions more rapidly, e.g. `wgmma` and new instructions from Blackwell (if any). This PR achieves this by describing the weight storage scheme using CUTLASS and CUTE.

Marlin doesn't support cutlass epilogues, so upstream is investigating subbyte weight quantization + activation quantization. 

### Benchmarks

1x H100, `hugging-quants/Meta-Llama-3.1-8B-Instruct-GPTQ-INT4`, `input_len=3072`, `output_len=512`

New kernels (machete):
```
Processed prompts: 100%|██████████| 1000/1000 [05:10<00:00,  3.22it/s, est. speed input: 9898.25 toks/s, output: 1650.25 toks/s]
Throughput: 3.21 requests/s, 11489.67 tokens/s
```

Old kernels (marlin):
```
Processed prompts: 100%|██████████| 1000/1000 [07:36<00:00,  2.19it/s, est. speed input: 6733.72 toks/s, output: 1122.65 toks/s]
Throughput: 2.18 requests/s, 7829.64 tokens/s
```